Create Fraudulent Activity Notifications.cpp
This problem is one of Hacker Rank's best conceptual problems, Conveying the importance of time complexity, O(n*log(n)). 
If you just write a calling function over d times for n elements, the complexity becomes O(n*d), which is non acceptable for large inputs.
and the other thing to keep in mind is the Median*.
Some codes recommend to Implement a Counting-sort, but anyway you can solve this problem with some slight modifications.